Refactor domain to avoid importing from registry, cleaners
