+++ core/controllers/question_editor.py
@@ -26,7 +26,6 @@ from core.constants import constants
 from core.controllers import acl_decorators
 from core.controllers import base
 from core.domain import fs_services
-from core.domain import html_cleaner
 from core.domain import image_validation_services
 from core.domain import question_domain
 from core.domain import question_services
@@ -111,7 +110,7 @@ class QuestionCreationHandler(base.BaseHandler):
             skill_difficulties)
         html_list = question.question_state_data.get_all_html_content_strings()
         filenames = (
-            html_cleaner.get_image_filenames_from_html_strings(html_list))
+            utils.get_image_filenames_from_html_strings(html_list))
         image_validation_error_message_suffix = (
             'Please go to the question editor for question with id %s and edit '
             'the image.' % question.id)
