+++ core/domain/story_domain.py
@@ -29,7 +29,6 @@ from core.domain import change_domain
 from core.domain import fs_domain
 
 from core.domain import fs_services  # pylint: disable=invalid-import-from # isort:skip
-from core.domain import html_cleaner  # pylint: disable=invalid-import-from # isort:skip
 from core.domain import html_validation_service  # pylint: disable=invalid-import-from # isort:skip
 
 # TODO(#14537): Refactor this file and remove imports marked
@@ -221,7 +220,7 @@ class StoryNode:
         self.destination_node_ids = destination_node_ids
         self.acquired_skill_ids = acquired_skill_ids
         self.prerequisite_skill_ids = prerequisite_skill_ids
-        self.outline = html_cleaner.clean(outline)
+        self.outline = utils.clean(outline)
         self.outline_is_finalized = outline_is_finalized
         self.exploration_id = exploration_id
 
@@ -686,7 +685,7 @@ class Story:
         self.thumbnail_bg_color = thumbnail_bg_color
         self.thumbnail_size_in_bytes = thumbnail_size_in_bytes
         self.description = description
-        self.notes = html_cleaner.clean(notes)
+        self.notes = utils.clean(notes)
         self.story_contents = story_contents
         self.story_contents_schema_version = story_contents_schema_version
         self.language_code = language_code
