+++ scripts/backend_test_shards.json
@@ -189,7 +189,6 @@
         "scripts.install_chrome_for_ci_test",
         "scripts.run_mypy_checks_test",
         "scripts.linters.pylint_extensions_test",
-        "core.domain.html_cleaner_test",
         "core.platform.taskqueue.dev_mode_taskqueue_services_test",
         "scripts.create_topological_sort_of_all_services_test",
         "core.storage.suggestion.gae_models_test",
