+++ scripts/run_mypy_checks.py
@@ -79,8 +79,6 @@ NOT_FULLY_COVERED_FILES = [
     'core/domain/fs_domain_test.py',
     'core/domain/fs_services.py',
     'core/domain/fs_services_test.py',
-    'core/domain/html_cleaner.py',
-    'core/domain/html_cleaner_test.py',
     'core/domain/html_validation_service.py',
     'core/domain/html_validation_service_test.py',
     'core/domain/image_validation_services.py',
